image_pair/requirements.txt
numpy
diffusers
transformers
matplotlib
scipy
cvxpy
tqdm
gdown
pillow
imageio
deepdiff
jupyterlab
notebook
opencv-python
opt-einsum
pandas
pyyaml
scikit-learn
seaborn
termcolor
streamlit
gradio
wandb
nvitop
lpips
git+https://github.com/openai/CLIP.git